+++ src/MGMParams.cpp
@@ -49,7 +49,7 @@ MGMParams::MGMParams(arma::vec& vec, int p, int ltot) {
     std::vector<arma::uword> lenSums(lens.size());
     lenSums[0] = lens[0];
 
-    for (int i = 1; i < lenSums.size(); i++) {
+    for (int i = 1; i < (int) lenSums.size(); i++) {
         lenSums[i] = lens[i] + lenSums[i-1];
     }
 
@@ -73,7 +73,7 @@ arma::vec MGMParams::toMatrix1D() {
     std::vector<int> lens{p*p, p, p*ltot, ltot*ltot, p, ltot};
     std::vector<arma::uword> lenSums(lens.size());
     lenSums[0] = lens[0];
-    for (int i = 1; i < lenSums.size(); i++) {
+    for (int i = 1; i < (int) lenSums.size(); i++) {
         lenSums[i] = lens[i] + lenSums[i-1];
     }
